Work on test to generate keys, continue on with encryption methods and tests
