+++ test/enigma_test.rb
@@ -9,9 +9,9 @@ class EnigmaTest < Minitest::Test
     assert_instance_of Enigma, enigma
   end
 
-  def test_enigma_can_generate_key
+  def test_enigma_can_generate_random_number
     enigma = Enigma.new
-    expected = enigma.generate_key
+    expected = enigma.generate_random_number
     assert_equal 5, expected.length
     assert_equal String, expected.class
   end
@@ -19,6 +19,20 @@ class EnigmaTest < Minitest::Test
   def test_convert_current_date_to_correct_format
     enigma = Enigma.new
     assert_equal "011521", enigma.generate_date
+  end
+
+  def test_to_generate_offsets
+    enigma = Enigma.new
+    assert_equal "1025", enigma.generate_offsets("040895")
+  end
+
+  def test_to_generate_keys
+    enigma = Enigma.new
+    enigma.generate_keys("02715", "1025")
+    assert_equal 3, enigma.a_key
+    assert_equal 27, enigma.a_key
+    assert_equal 73, enigma.a_key
+    assert_equal 20, enigma.a_key
   end
 
   def test_enigma_can_encrypt
